Issues #44 and #46-#50: Add complete inventory seed data

+++ src/ProjectHopeLynden.Application/Inventory/InventoryEntryListItem.cs
@@ -4,7 +4,7 @@ public sealed record InventoryEntryListItem(
     int Id,
     string ItemName,
     string LocationName,
-    int CurrentQuantity,
+    double CurrentQuantity,
     bool IsCommodity,
     DateTime? BestByDate,
     bool IsMenuItem,
